refactor: rename snippet loading service
