+++ src/extension.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import { addHttpSnippet } from "./add-http-snippet";
+import { addHttpSnippet } from "./add-snippet";
 
 export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
